+++ sample/samplez.rb
@@ -2,17 +2,19 @@ require 'num4mechaequ'
 
 class Num4MechaEquTest
     def initialize
-        @k = 1.0
-        @m = 1.0
-        @h0 = 0.0
-        @v0 = 1.0
+        @k = 1.0                       # バネ定数
+        @c = 0.0                       # 空気抵抗
+        @m = 1.0                       # 重りの重さ
+        @h0 = 0.0                      # 初期位置
+        @v0 = 1.0                      # 初期速度
+        @theta = 5 * Math::PI / 180.0  # 角度
     end
     #
     # 自由落下による運動方程式テスト
-    def motiEqu4freeFallEquTest
-        puts 'motiEqu4freeFallTest in'
+    def freeFallMotionTest
+        puts 'freeFallMotionTest in'
         yi_1 = []
-        yi_1 = Num4MechaEquLib.motiEqu4freeFallEqu(@m, @h0, @v0)
+        yi_1 = Num4MechaEquLib.freeFallMotion(@m, @c, 2, 0.0, 0.0)
         yi_1.map{ |v|
             print v
             puts
@@ -20,11 +22,24 @@ class Num4MechaEquTest
         puts
     end
     #
-    # バネの固有振動テスト
-    def springFreqEquTest
-        puts 'springFreqEquTest in'
+    # 放物運動
+    def projectileMotionTest
+        puts 'projectileMotionTest in'
         yi_1 = []
-        yi_1 = Num4MechaEquLib.springFreqEqu(@k, @m, 1, 0)
+        yi_1 = Num4MechaEquLib.projectileMotion(@m, @theta, 1, 0.0, 3.0)
+        yi_1.map{ |v|
+            print v
+            puts
+        }
+        puts
+    end
+
+    #
+    # 単振動テスト
+    def SHMTest
+        puts 'SHMTest in'
+        yi_1 = []
+        yi_1 = Num4MechaEquLib.SHM(@k, @m, 1, 1, 0)
         yi_1.map{ |v|
             print v
             puts
@@ -33,6 +48,7 @@ class Num4MechaEquTest
     end
 end
 tst = Num4MechaEquTest.new
-tst.springFreqEquTest()
-tst.motiEqu4freeFallEquTest()
+tst.SHMTest()
+tst.freeFallMotionTest()
+tst.projectileMotionTest()